# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MAROC EXPLORER API",
        "description": "This is the API documentation for the MAROC EXPLORER API. It handles webhook requests and responses, allowing users to generate and display data dynamically.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/itineraires": {
            "get": {
                "tags": [
                    "Itineraires"
                ],
                "summary": "Display a listing of all itineraires",
                "description": "Returns a list of all itineraires with their associated data.",
                "operationId": "6b60255d8d64f2182046d68efa480ff9",
                "responses": {
                    "200": {
                        "description": "Itineraires fetched successfully."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Itineraires"
                ],
                "summary": "Store a newly created itineraire",
                "description": "Creates a new itineraire and returns the created record.",
                "operationId": "ba1645a770c7ac4c8fd9b7ab28363677",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "duration",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Exploring Marrakech"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Cultural"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "3 days"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "marrakech.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Itineraire created successfully."
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/itineraires/{id}": {
            "get": {
                "tags": [
                    "Itineraires"
                ],
                "summary": "Display a specific itineraire",
                "description": "Returns a specific itineraire with its associated data.",
                "operationId": "bfeb5668e786cc44fa25592da5bcb544",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the itineraire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itineraire fetched successfully."
                    },
                    "404": {
                        "description": "Itineraire not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Itineraires"
                ],
                "summary": "Update the specified itineraire",
                "description": "Updates a specific itineraire and returns the updated record.",
                "operationId": "3a952bd0498de6b327ead7520ecf7acb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the itineraire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Updated Category"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "4 days"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "updated.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Itineraire updated successfully."
                    },
                    "403": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Itineraire not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Itineraires"
                ],
                "summary": "Remove the specified itineraire",
                "description": "Deletes a specific itineraire.",
                "operationId": "d7b1290f0756075ad8db19d63dc49505",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the itineraire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itineraire deleted successfully."
                    },
                    "403": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Itineraire not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/itineraires/search/{query}": {
            "get": {
                "tags": [
                    "Itineraires"
                ],
                "summary": "Search itineraires",
                "description": "Search itineraires by title or category or duration.",
                "operationId": "f7a1401845c5be02d586b776aff5db5b",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "The search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itineraires found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}